Add suggestions from review.
