+++ src/dms/core/EdFi.DataManagementService.Core/OAuth/IOAuthManager.cs
@@ -1,0 +1,18 @@
+// SPDX-License-Identifier: Apache-2.0
+// Licensed to the Ed-Fi Alliance under one or more agreements.
+// The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
+// See the LICENSE and NOTICES files in the project root for more information.
+
+using EdFi.DataManagementService.Core.External.Model;
+
+namespace EdFi.DataManagementService.Core.OAuth;
+
+public interface IOAuthManager
+{
+    public Task<HttpResponseMessage> GetAccessTokenAsync(
+        IHttpClientWrapper httpClient,
+        string authHeaderString,
+        string upstreamUri,
+        TraceId traceId
+    );
+}
